requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.12
langchain-openai>=0.0.5
python-dotenv
fastapi
uvicorn
gradio
httpx
faiss-cpu